Fix: Slider default step causes crash on Windows OS (#260)
* Calculate stepFrequency for default step

The default value of step is set to zero for all platforms.
On Windows OS this approach causes the slider XAML resolution to be at
the level of infinite. Further it causes the application to crash
whenever user tries to move the slider.

To fix that this commit implements the step calculation done when
noticing the default value of <Slider/> to be zero (or zero set by user)
and return the percentage value of Slider's range (from min to max).

* Update README with default step for Windows

Due to changing the behavior of Slider's step for Windows the README
required an update. The description of default step approach mentiones
only the range calculation without the implementation details.

* Change min and max values to doubles

The integer type was previously used for the min and max values of the
Slider on Windows.
This was however changed in the delivery of:
https://github.com/callstack/react-native-slider/pull/259/files
but was partially inverted as the result of rebasing the conflicting
changes.
As the int->double change is good to have and should be kept, this
commit fixes the merge mistake by changing the integer type to double.